Added a dry run option

+++ relink.py
@@ -12,11 +12,12 @@ def get_args(args:T.List[str]) -> argparse.Namespace:
     Parse command line arguments
     """
     parser = argparse.ArgumentParser()
-    parser.add_argument("--new-root", default=Path.cwd())
+    parser.add_argument("--dry-run", action="store_true")
+    parser.add_argument("--new-root", default=Path.cwd(), type=Path)
     parser.add_argument("--old-root", action="extend", nargs="+", required=True, type=Path)
 
     parsed = parser.parse_args(args)
-    parsed.old_root = set(parsed.old_root)
+    parsed.old_root = set(parsed.old_root)  # Remove duplicates
 
     return parsed
 
@@ -38,6 +39,11 @@ def read_lines(f=sys.stdin, delimiter:str = "\0") -> T.Iterator[str]:
         yield from (previous, *tail[:-1])
         previous = tail[-1]
 
+    # This should never happen, because a properly delimited input
+    # should have a trailing delimiter...but just in case
+    if previous != "":
+        yield previous
+
 
 if __name__ == "__main__":
     args = get_args(sys.argv[1:])
@@ -54,13 +60,13 @@ if __name__ == "__main__":
                 new_target_rel = Path(os.path.relpath(new_target_abs, link.parent))
 
                 # Unlink, relink and output (link, old target, new target)
-                link.unlink()
-                link.symlink_to(new_target_rel)
                 print(f"{link}\t{target}\t{new_target_rel}")
+                if not args.dry_run:
+                    link.unlink()
+                    link.symlink_to(new_target_rel)
 
                 break
 
         else:
             # Not an in-tree link (output to stderr)
             print(f"{link}", file=sys.stderr)
-            continue
